Refactor dropMigrationProcessConfiguration for common usage
